pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docbrief"
version = "0.1.0"
description = "Convert a folder of raw documents into an AI-optimized briefing packet"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.23",
    "python-docx>=1.1",
]

[project.scripts]
docbrief = "docbrief.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["docbrief*"]
